Create directory and page for bidder side. Move some files
