+++ client/src/components/ui/on-boarding.tsx
