+++ client/src/pages/index.tsx
@@ -1,13 +1,13 @@
 import { Inter as FontSans } from "next/font/google";
 import { useEffect, useState } from "react";
 
-import OnBoarding from "@/components/on-boarding";
+import OnBoarding from "@/components/ui/on-boarding";
 import { usePings } from "@/hooks/pings";
 import { cn } from "@/lib/utils";
 
-import TaskDetails from "../components/task-details";
 import { Button } from "../components/ui/button";
 import SplashScreen from "../components/ui/splash-screen";
+import TaskDetails from "../components/ui/task-details";
 
 const fontSans = FontSans({
   subsets: ["latin"],
